+++ Example/Tests/Tests.m
@@ -16,11 +16,11 @@ SpecBegin(InitialSpecs)
 describe(@"SEGCleverTapIntegration", ^{
     describe(@"initWithSettings", ^{
         it(@"initializes CleverTap if accountID and account Token are set", ^{
-            NSDictionary *settings =  @{@"CleverTapAccountID":CleverTapAccountID, @"CleverTapToken":CleverTapAccountToken};
+            //NSDictionary *settings =  @{@"CleverTapAccountID":CleverTapAccountID, @"CleverTapToken":CleverTapAccountToken};
             id cleverTapMock = OCMClassMock([CleverTap class]);
             OCMExpect([cleverTapMock changeCredentialsWithAccountID:CleverTapAccountID andToken:CleverTapAccountToken]);
             OCMExpect([[cleverTapMock sharedInstance] notifyApplicationLaunchedWithOptions:nil]);
-            SEGCleverTapIntegration *cleverTapIntegration = [[SEGCleverTapIntegration alloc] initWithSettings:settings];
+            //SEGCleverTapIntegration *cleverTapIntegration = [[SEGCleverTapIntegration alloc] initWithSettings:settings];
             OCMVerifyAllWithDelay(cleverTapMock, 2);
         });
     });
@@ -62,7 +62,8 @@ describe(@"SEGCleverTapIntegration", ^{
                                      @"phone" : @"5555555555"
                                      };
             
-            SEGIdentifyPayload *identifyPayload = [[SEGIdentifyPayload alloc] initWithUserId:@"testUser" anonymousId:nil traits:traits context:nil integrations:nil];
+            
+            SEGIdentifyPayload *identifyPayload = [[SEGIdentifyPayload alloc] initWithUserId:@"testUser" anonymousId:@"123456" traits:traits context:@{} integrations:@{}];
             
             [cleverTapIntegration identify:identifyPayload];
             
@@ -88,8 +89,8 @@ describe(@"SEGCleverTapIntegration", ^{
             
             SEGTrackPayload *trackPayload = [[SEGTrackPayload alloc] initWithEvent:@"testEvent"
                                                                         properties:properties
-                                                                           context:nil
-                                                                      integrations:nil];
+                                                                           context:@{}
+                                                                      integrations:@{}];
             [cleverTapIntegration track:trackPayload];
             OCMVerifyAllWithDelay(cleverTapMock, 2);
         });