+++ ExampleSwift/Pods/Analytics/Analytics/Classes/Internal/SEGHTTPClient.m
@@ -12,6 +12,14 @@
     };
 }
 
++ (NSString *)authorizationHeader:(NSString *)writeKey
+{
+    NSString *rawHeader = [writeKey stringByAppendingString:@":"];
+    NSData *userPasswordData = [rawHeader dataUsingEncoding:NSUTF8StringEncoding];
+    return [userPasswordData base64EncodedStringWithOptions:0];
+}
+
+
 - (instancetype)initWithRequestFactory:(SEGRequestFactory)requestFactory
 {
     if (self = [self init]) {
@@ -24,26 +32,25 @@
     return self;
 }
 
-- (NSString *)authorizationHeader:(NSString *)writeKey
-{
-    NSString *rawHeader = [writeKey stringByAppendingString:@":"];
-    NSData *userPasswordData = [rawHeader dataUsingEncoding:NSUTF8StringEncoding];
-    return [userPasswordData base64EncodedStringWithOptions:0];
-}
 
 - (NSURLSessionUploadTask *)upload:(NSDictionary *)batch forWriteKey:(NSString *)writeKey completionHandler:(void (^)(BOOL retry))completionHandler
 {
+    //    batch = SEGCoerceDictionary(batch);
     NSURLSessionConfiguration *config = [NSURLSessionConfiguration defaultSessionConfiguration];
     config.HTTPAdditionalHeaders = @{
         @"Accept-Encoding" : @"gzip",
         @"Content-Encoding" : @"gzip",
         @"Content-Type" : @"application/json",
-        @"Authorization" : [@"Basic " stringByAppendingString:[self authorizationHeader:writeKey]],
+        @"Authorization" : [@"Basic " stringByAppendingString:[[self class] authorizationHeader:writeKey]],
     };
     NSURLSession *session = [NSURLSession sessionWithConfiguration:config];
 
-    NSURL *url = [NSURL URLWithString:@"https://api.segment.io/v1/batch"];
+    NSURL *url = [SEGMENT_API_BASE URLByAppendingPathComponent:@"batch"];
     NSMutableURLRequest *request = self.requestFactory(url);
+    
+    // This is a workaround for an IOS 8.3 bug that causes Content-Type to be incorrectly set
+    [request addValue:@"application/json" forHTTPHeaderField:@"Content-Type"];
+    
     [request setHTTPMethod:@"POST"];
 
     NSError *error = nil;
@@ -93,10 +100,11 @@
         completionHandler(YES);
     }];
     [task resume];
+    [session finishTasksAndInvalidate];
     return task;
 }
 
-- (NSURLSessionDataTask *)settingsForWriteKey:(NSString *)writeKey completionHandler:(void (^)(BOOL success, NSDictionary *settings))completionHandler
+- (NSURLSessionDataTask *)settingsForWriteKey:(NSString *)writeKey completionHandler:(void (^)(BOOL success, JSON_DICT _Nullable settings))completionHandler
 {
     NSURLSessionConfiguration *config = [NSURLSessionConfiguration defaultSessionConfiguration];
     config.HTTPAdditionalHeaders = @{
@@ -104,8 +112,7 @@
     };
     NSURLSession *session = [NSURLSession sessionWithConfiguration:config];
 
-    NSString *rawURL = [NSString stringWithFormat:@"https://cdn-settings.segment.com/v1/projects/%@/settings", writeKey];
-    NSURL *url = [NSURL URLWithString:rawURL];
+    NSURL *url = [SEGMENT_CDN_BASE URLByAppendingPathComponent:[NSString stringWithFormat:@"/projects/%@/settings", writeKey]];
     NSMutableURLRequest *request = self.requestFactory(url);
     [request setHTTPMethod:@"GET"];
 
@@ -134,10 +141,11 @@
         completionHandler(YES, responseJson);
     }];
     [task resume];
+    [session finishTasksAndInvalidate];
     return task;
 }
 
-- (NSURLSessionDataTask *)attributionWithWriteKey:(NSString *)writeKey forDevice:(NSDictionary *)context completionHandler:(void (^)(BOOL success, NSDictionary *properties))completionHandler;
+- (NSURLSessionDataTask *)attributionWithWriteKey:(NSString *)writeKey forDevice:(JSON_DICT)context completionHandler:(void (^)(BOOL success, JSON_DICT _Nullable properties))completionHandler;
 
 {
     NSURLSessionConfiguration *config = [NSURLSessionConfiguration defaultSessionConfiguration];
@@ -145,11 +153,11 @@
         @"Accept-Encoding" : @"gzip",
         @"Content-Encoding" : @"gzip",
         @"Content-Type" : @"application/json",
-        @"Authorization" : [@"Basic " stringByAppendingString:[self authorizationHeader:writeKey]],
+        @"Authorization" : [@"Basic " stringByAppendingString:[[self class] authorizationHeader:writeKey]],
     };
     NSURLSession *session = [NSURLSession sessionWithConfiguration:config];
 
-    NSURL *url = [NSURL URLWithString:@"https://mobile-service.segment.com/v1/attribution"];
+    NSURL *url = [MOBILE_SERVICE_BASE URLByAppendingPathComponent:@"/attribution"];
     NSMutableURLRequest *request = self.requestFactory(url);
     [request setHTTPMethod:@"POST"];
 
@@ -194,6 +202,7 @@
         completionHandler(YES, responseJson);
     }];
     [task resume];
+    [session finishTasksAndInvalidate];
     return task;
 }