+++ ExampleSwift/Pods/Target Support Files/Analytics/Analytics-umbrella.h
@@ -1,5 +1,13 @@
 #ifdef __OBJC__
 #import <UIKit/UIKit.h>
+#else
+#ifndef FOUNDATION_EXPORT
+#if defined(__cplusplus)
+#define FOUNDATION_EXPORT extern "C"
+#else
+#define FOUNDATION_EXPORT extern
+#endif
+#endif
 #endif
 
 #import "SEGAES256Crypto.h"
@@ -9,6 +17,7 @@
 #import "SEGIdentifyPayload.h"
 #import "SEGIntegration.h"
 #import "SEGIntegrationFactory.h"
+#import "SEGIntegrationsManager.h"
 #import "SEGPayload.h"
 #import "SEGScreenPayload.h"
 #import "SEGTrackPayload.h"
@@ -16,7 +25,6 @@
 #import "SEGAnalyticsUtils.h"
 #import "SEGFileStorage.h"
 #import "SEGHTTPClient.h"
-#import "SEGReachability.h"
 #import "SEGSegmentIntegration.h"
 #import "SEGSegmentIntegrationFactory.h"
 #import "SEGStorage.h"
@@ -24,7 +32,12 @@
 #import "SEGUserDefaultsStorage.h"
 #import "SEGUtils.h"
 #import "UIViewController+SEGScreen.h"
+#import "SEGContext.h"
+#import "SEGMiddleware.h"
 #import "SEGAnalytics.h"
+#import "SEGAnalyticsConfiguration.h"
+#import "SEGSerializableValue.h"
+#import "SEGReachability.h"
 
 FOUNDATION_EXPORT double AnalyticsVersionNumber;
 FOUNDATION_EXPORT const unsigned char AnalyticsVersionString[];